now its not so much code, so put pyproject stuff in one file
