+++ src/somesy/pyproject/__init__.py
@@ -1,4 +1,4 @@
 """Pyproject module."""
-from .poetry import Poetry
+from .writer import Poetry, Pyproject, SetupTools
 
-__all__ = ["Poetry"]
+__all__ = ["Pyproject", "Poetry", "SetupTools"]
